% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jitter_dt.R
\name{create_jitter_dt}
\alias{create_jitter_dt}
\title{Create Jittered Date Table}
\usage{
create_jitter_dt(dataset, keycol_name, max_days = 100, seed = 123)
}
\arguments{
\item{dataset}{The input data.frame that contains a column which is the key (unique identifier)}

\item{keycol_name}{A string that is the name of the column that the unique id is located in}

\item{max_days}{An integer representing the largest number of days that the dates can be jittered}

\item{seed}{The random seed used to generate the dataset. This can be used to reconstruct the id mapping at a later stage.}
}
\value{
A data.frame that contains 2 columns named original.ID and dates to jitter
}
\description{
This function creates a lookup dataframe where one column is the original key of an input dataframe and the other is a random number of days (by patient) 
to jitter the dates
}
