% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookup.R
\name{create_lookup}
\alias{create_lookup}
\title{Create ID Lookup Table}
\usage{
create_lookup(dataset, keycol_name, digits = 7, seed = 123)
}
\arguments{
\item{dataset}{The input data.frame that contains a column which is the key (unique identifier)}

\item{keycol_name}{A string that is the name of the column that the unique id is located in}

\item{digits}{An integer number of digits of the replacement id. The default is 7 to allow for enough entries. Note that increasing the argument too much will result in a large amount of memory used to internally store the range of all possible values.}

\item{seed}{The random seed used to generate the dataset. This can be used to reconstruct the id mapping at a later stage.}
}
\value{
A data.frame that contains 2 columns named original.ID and replacement.ID
}
\description{
This function creates a lookup dataframe where one column is the original key of an input dataframe and the other is the replacement ID
}
