% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_dates.R
\name{jitter_dates}
\alias{jitter_dates}
\title{Jitter Dates}
\usage{
jitter_dates(dataset, keycol_name, date_cols, direction = "backward", lookup)
}
\arguments{
\item{dataset}{The input data.frame that contains a column which is the key (unique identifier)}

\item{keycol_name}{A string that is the name of the column that the unique id is located in dataset}

\item{date_cols}{A named vector of strings where the elements are the names of the columns in dataset that are presumed dates. The names of the elements are the lubridate-esque formats of the dates, e.g. "mdy_hms"}

\item{direction}{Should the function jitter dates forward or backward? Defaults to "backward"}

\item{lookup}{lookup is a 2-column dataframe with the first column being the key and the second column being the number of days to jitter. Usually the output from create_jitter_dt}
}
\value{
The original dataframe with all dates jittered
}
\description{
This function creates a lookup dataframe where one column is the original key of an input dataframe and the other is a random number of days (by patient) 
to jitter the dates
}
