% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_idcol.R
\name{replace_idcol}
\alias{replace_idcol}
\title{Replace ID column}
\usage{
replace_idcol(dataset, col_toreplace, lookup)
}
\arguments{
\item{dataset}{The input data.frame that contains the id to be replaced}

\item{col_toreplace}{A string that is the name of the column to be replaced}

\item{lookup}{The lookup 2-column data.frame that contains the original column as the first column and the corresponding IDs as the second. Ideally generated by \code{"create_lookup"}}
}
\value{
A data.frame which is like the original except that the ID column is replaced with its corresponding entry in the lookup table
}
\description{
This function replaces an ID column in a dataframe with another ID, given a lookup table
}
